test: isolate failure session execution coverage
